+++ bcbio/variation/validate.py
@@ -101,9 +101,9 @@ def _create_validate_config(vrn_file, rm_file, rm_interval_file, rm_genome,
            "calls": [ref_call, eval_call]}
     if data.get("callable_bam"):
         exp["align"] = data["callable_bam"]
-        intervals = ensemble.get_analysis_intervals(data)
-        if intervals:
-            exp["intervals"] = os.path.abspath(intervals)
+    intervals = ensemble.get_analysis_intervals(data)
+    if intervals:
+        exp["intervals"] = os.path.abspath(intervals)
     return {"dir": {"base": base_dir, "out": "work", "prep": "work/prep"},
             "experiments": [exp]}