+++ HISTORY.md
@@ -4,6 +4,8 @@
   files) and RNA-seq (transcript GTF files) analyses into an organism specific
   resources file. Improves ability to support additional organisms and genome
   builds.
+- Pull `description` sample names from BAM files if not present in input
+  configuration file. Thanks to Paul Tang for suggestion.
 
 ## 0.7.2 (August 30, 2013)